+++ old_tests/test_parse.py
@@ -42,7 +42,7 @@ class TestParse:
 
     def test_extract(self, parsed_mscx):
         target = self.test_results
-        path_dict = parsed_mscx.output_dataframes(measures_folder=target, measures_suffix="_measures",
+        path_dict = parsed_mscx.store_extracted_facets(measures_folder=target, measures_suffix="_measures",
                                                   notes_folder=target, notes_suffix='_notes',
                                                   labels_folder=target, labels_suffix='_labels')
         # for path, what in path_dict.items():